backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
datasets==2.10.1
dill==0.3.6
distlib==0.3.6
dj-database-url==1.2.0
Django>=3.0
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
dulwich==0.21.3
filelock==3.10.0
frozenlist==1.3.3
fsspec>=2023.1.0
gunicorn==20.1.0
html5lib==1.1
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
networkx>=2.6.3
numpy>=1.21.6
packaging==23.0
pandas>=1.3.5
pexpect==4.8.0
Pillow==9.4.0
pipenv==2023.2.18
piq==0.7.0
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
psycopg2==2.9.5
ptyprocess==0.7.0
pyarrow==11.0.0
pycparser==2.21
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv>=0.20.0
pytz==2022.7.1
PyYAML==6.0
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-toolbelt==0.10.1
responses==0.18.0
shellingham==1.5.0.post1
six==1.16.0
sqlparse==0.4.3
sympy>=1.10.1
tokenizers==0.13.2
tomlkit==0.11.6
torch>=1.13.1
torchvision>=0.14.1
tqdm==4.65.0
transformers==4.27.2
trove-classifiers==2023.3.9
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
virtualenv-clone==0.5.7
webencodings==0.5.1
xattr==0.10.1
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0